CoversionBase v1 and Database Fixes
Can now add/remove/edit need data in the client and properly converse with the server.
Users will update the MapTag closest to them.
If the provided food amount is not enough, the MapTag will not dissapear and instead lower its AnimalCount.
Maptags are automatically updated from the server on certain events, but there is also a button to test manually too.
**Has a few minor bugs.